Refactored for layers and prod release
