--- lambda_layers/mindtether_core/__init__.py
@@ -1,14 +1,0 @@
-import os
-
-class MindTetherResources:
-    
-    
-    
-    def __init__(self) -> None:
-        pass
-
-    def Asset():
-        def FONT():
-            THE_BOLD_FONT = "fonts/theboldfont.ttf"    
-    
-
