+++ lambda_layers/mindtether_core/python/MindTetherCore/__init__.py
@@ -1,0 +1,13 @@
+from os import path
+
+
+    
+    
+    # def __init__(self) -> None:
+    #     pass
+
+
+_font_dir = path.join(path.dirname(__file__),"fonts")
+FONT = {
+    "THE_BOLD_FONT" : _font_dir + "theboldfont.ttf"
+}
